+++ EssentialFeed/EssentialFeed/Feed Cache/FeedStore.swift
@@ -1,0 +1,17 @@
+//
+//  FeedStore.swift
+//  EssentialFeed
+//
+//  Created by Cristian Felipe Patiño Rojas on 3/4/25.
+//
+
+
+import Foundation
+
+public protocol FeedStore {
+    typealias DeletionCompletion = (Error?) -> Void
+    typealias InsertionCompletion = (Error?) -> Void
+    
+    func deleteCachedFeed(completion: @escaping DeletionCompletion)
+    func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
+}
